+++ lib/use_cases.rb
@@ -20,10 +20,19 @@ module UseCases
 
   def self.configure(&blk)
     super
-    subscribe
+    finalize_subscriptions!
   end
 
-  def self.subscribe
+  def self.finalize_subscriptions!
+    subscribe(*subscribers)
+    return unless UseCases::Events::Subscriber.respond_to?(:descendants)
+
+    usecase_subscriber_children = UseCases::Events::Subscriber.descendants
+    subscribe(*usecase_subscriber_children)
+    subscribers.concat(usecase_subscriber_children)
+  end
+
+  def self.subscribe(*subscribers)
     subscribers.each { |subscriber| publisher.subscribe(subscriber) }
   end
 end