+++ lib/use_cases/events/publish_job.rb
@@ -5,6 +5,7 @@ module UseCases
   module Events
     class PublishJob < ActiveJob::Base
       def perform(publish_key, payload)
+        publish_key += '.async'
         UseCases.publisher.publish(publish_key, payload)
       end