+++ lib/use_cases/events/subscriber.rb
@@ -1,0 +1,6 @@
+module UseCases
+  module Events
+    class Subscriber
+    end
+  end
+end